requirements.txt
Flask~=3.1.0
geopandas~=1.0.1
pandas~=2.2.3
pyproj~=3.7.0
shapely~=2.0.6
python-dotenv~=1.0.1
openpyxl~=3.1.5
watchdog~=6.0.0
PyYAML~=6.0.2